R Code
This SAS script was provided by Dr. Pianpian Cao as part of a lung cancer screening project. It serves as a reference for recoding variables from the IPUMS NHIS dataset (2010–2023) into formats required for use in the LC Risk microsimulation model. Key transformations include:

-Recoding demographic variables (age, sex, race, education)
-Cleaning and calculating smoking history (e.g., smoking duration, years since quitting)
-Creating binary indicators for comorbidities (COPD, diabetes, cancer, etc.)
-Constructing family history indicators from multiple lung cancer-related variables

This file is intended as a reference for translating logic into R for future reproducible analyses.